Complete rewrite, stricter OOP principles, add multiple teams per user.
This is a complete architectural rewrite. It breaks the plugin up into
components, and defines a class for each one. The classes are stricter
about maintaining a SOLID OOP design philosophy.

Additionally, a new custom post type is defined for a response team and
so a single user can now have multiple response teams. An alert can be
sent to at least one but possibly multiple response teams. The first
team a user creates becomes their default team unless they set their
default team to be a different team.