Initial experiment with "built-in" WordPress chat using post comments.
This commit begins an experimental branch in which the WordPress
commenting system is used as the back-end for a "chat room" feature.
Individual alerts will remember the chat system they are using, and an
admin will be able to set a default chat system for new alerts. This
maintains the possibility for letting users choose their own chat system
independent of the administrator or on a per-alert basis (although that
seems like overkill).

If this proves promising enough, the hope is that the built-in chat room
will be themable in the same way as the WordPress front-end commenting
templates are, allowing further customization of the plugin's interface.

git-svn-id: https://plugins.svn.wordpress.org/buoy/trunk@1328767 b8457f37-d9ea-0310-8a92-e5e31aec5664