Closes #111: Use a separate plugin for outgoing GPG encrypted email.
Revert "Closes #39: Encrypt outgoing alerts when user GPG public key exists."

This reverts commit 05b284c01405e94795be8480b963475a5203004d.

I have written an entirely separate plugin for WordPress to handle all
outgoing emails with PGP encryption. That plugin is the proper place for
this feature set, not in Buoy. A Buoy admin will need to add that other
plugin, available under the name "WP PGP Encrypted Emails," to their
Buoy-enabled site, but we can mention this fact in our security advice
wiki page. The benefits of this modularization far outweigh the costs of
requiring an additional plugin to be installed by the admin, though.