Fixes #119: Check for missing function declaration.
Some sites exhibit a fatal error when loading the Theme Customizer when
Buoy is enabled. I don't know why this is because my own system does not
reproduce this issue. Nevertheless, the fix seems to be to check for the
presence of the function and, if not, to `require` it ourselves. This is
easy enough, if annoying, so we'll work around this for the time being
until we have more time to figure out why this behavior doesn't seem to
be consistent.

Since I can't reliably reproduce this, I also can't write tests for it.
Sad panda. :(