Fixes #121: Usernames with spaces in them were modified unexpectedly.
The `user_nicename` is not the same thing as the `user_login` property
for `WP_User` objects. Using the former converts spaces to dashes, which
is not what Buoy's input expected. Using the `user_login` property is
what Buoy expects to receive, so it is what we should output, too.