Remove Google Maps and replace with OpenStreetMaps backed Leaflet Map.
This is important for a few reasons, not least of which are:

* Better privacy
* Less reliance on third-party scripts
* More potential future features

Better privacy because the act of pulling down Google Maps' location
tiles tells Google where a given user is in the real world, leaking
their location to a company that doesn't need to have more private data.

Less reliance on third-party scripts because we bundle the Leaflet JS in
the plugin itself so we no longer rely on receiving Google's JavaScript
files from their system.

More potential future features because, as discussed in issue #62 on our
GitHub, the Google Maps API terms of use prohibit offline caching of map
tiles. We want to be able to provide offline mapping functionality at
some point in the future, so we would need to use a free software map
tool to do this, anyway. Now, we are, and can build with that in mind.

git-svn-id: https://plugins.svn.wordpress.org/buoy/trunk@1389133 b8457f37-d9ea-0310-8a92-e5e31aec5664