Increase SSE connection restart frequency to 1 minute.
This is needed for the (common?) situation in which a given responder
sends more than one chat message in quick succession. The server will
have kept the previous connection open for up to three minutes(!) prior
to this commit, which is probably going to drain the worker pools of
lower-end servers. This is probably a better balance between HTTP
requests and performance.