Hot fix for fatal error during first install caused by migration code.
I'm not sure why this wasn't caught by the unit tests, since AFAIK the
test harness tries activating the plugin from a clean database. :(

This should be safe because WordPress plugin repository reports all
activate versions of Buoy are at 0.2.x, and this code was intended for
migrating from 0.1.x to 0.2.x.