WordPress 4.7 core now includes the WP REST API's "Content API"!
We no longer require any plugin dependencies; the only one Buoy used to
need was the REST API plugin, but the parts we use are now shipped as
part of WP core. Hooray! See https://codex.wordpress.org/Version_4.7

This commit removes the defensive activation-time check for the REST API
and wraps the initialize-time check for the REST API's presence with a
`version_compare()` to ensure that only pre-4.7 versions are actually
forcibly installing this additional plugin.