Complete first pass of the UI conversion to Backbone.
This commit continues in the pattern from the previous commit, removing
the `buoy_dom_hooks` where they are no longer needed and converting each
individual user interface element to a `Backbone.View` object. Notably:

* Alert button names have been standardized across the codebase.
* The `attachHandlers` method was replaced with Backbone constructors.
* Bug in `class-buoy-alert.php` when scheduling Timed Alerts was fixed.
* JSDoc syntax is now used thoroughly across the `alert.js` file. Yay!

This isn't perfect yet. There are a few parts where the lack of a real
`Backbone.Model` for our alerts are requiring some hacky things. But the
purpose of this commit was to begin the migration to a Backbone-based
structure as gradually as is sensible to do so. Models will come later.