Prepare for an (aspirational) JavaScript testing environment.
This commit includes BackboneJS as part of our `package.json`'s
`devDependencies` and configures Karma to load it as necessary. This
means we are mostly ready for testing Backbone components when and if we
figure out what, exactly, we'd like to test.

Previous to this commit, there was a rather contrived test for a single
method. However, that was only ever being used in one place, so I've
refactored it and removed it from the test suite as it's a single jQuery
call, anyway.