Update to Bootstrap 3.3.7, refactor workaround code for Modal bug.
Sadly the Bootstrap Modal issue still isn't fixed in 3.3.7, if it indeed
a BootstrapCSS bug and not something else. Either way, we might as well
update to the latest release. But also, I want to future-proof our
workaround for this issue just in case we add more pages and forget it's
still a problem, so I refactored that workaround a smidge.