Fixes #207: Avoid info disclosure even if PHP is insecurely configured.
This commit addresses a potential information disclosure vulnerability
when PHP is configured to `display_errors` by suppressing the output of
PHP's warnings in the case where it would otherwise expose sensitive
information about the running system. This is done to protect the users
of a Buoy even in the face of a less experienced sysadmin who leaves
development configurations in effect on a production deployment.

Further, it enhances the mechanism by which the Timed Alert feature is
enabled away from a simplistic "is this a POSIX system?" guess and to
capability detection of whether or not the system has and can automate
its cron daemon. This logic has been concentrated in the crontab manager
class, which now throws a `RuntimeException` caught elsewhere. This also
enables the system to auto-correct its plugin settings, automatically
disabling the Timed Alert feature should a PHP-level configuration
change take place unbeknownst to the Buoy Admin.

Finally, the user interface of Buoy's Settings screen now explains why
or why not the feature is available on the running installation.

Old unit tests that were sort of but not really testing the earlier
implementation were removed. This is because the test is really an
integration test as it relies on the OS-level specifics and PHP's own
configuration values.